refactor: decompose func.rs into focused modules — 1144 lines → 10 files
Extract helpers (to_snake_case, to_pascal_case, is_string_expr),
statement conversion (convert_stmt, convert_stmt_recursive),
function declaration processing (process_fn_decl), and expression
handling into an expr/ module (binary, call, member, arrow, literal,
misc). All files under 400 lines. Delete func.rs after full migration.
