--- crates/tyrus_codegen/src/convert/func.rs
@@ -1,1144 +1,0 @@
-use quote::{format_ident, quote};
-use swc_ecma_ast::{
-    ArrowExpr, BinExpr, BinaryOp, BlockStmtOrExpr, CallExpr, Callee, Decl, Expr, ExprOrSpread,
-    FnDecl, Lit, MemberExpr, Pat, Stmt, UpdateExpr, UpdateOp,
-};
-
-use super::type_mapper::{map_ts_type, unwrap_promise_type};
-
-// Helper function to snake case (keep standalone)
-pub fn to_snake_case(str: &str) -> String {
-    let mut s = String::with_capacity(str.len());
-    let mut was_upper = false;
-    for (i, c) in str.chars().enumerate() {
-        if c.is_uppercase() {
-            if i > 0 && !was_upper {
-                s.push('_');
-            }
-            s.push(c.to_ascii_lowercase());
-            was_upper = true;
-        } else {
-            s.push(c);
-            was_upper = false;
-        }
-    }
-    s
-}
-
-// Helper to pascal case (keep standalone)
-pub fn to_pascal_case(str: &str) -> String {
-    let mut s = String::with_capacity(str.len());
-    let mut next_upper = true;
-    for c in str.chars() {
-        if c == '_' || c == '-' {
-            next_upper = true;
-        } else if next_upper {
-            s.push(c.to_ascii_uppercase());
-            next_upper = false;
-        } else {
-            s.push(c);
-        }
-    }
-    s
-}
-
-impl super::interface::RustGenerator {
-    pub fn process_fn_decl(&mut self, n: &FnDecl) {
-        let fn_name = to_snake_case(&n.ident.sym);
-        let fn_ident = format_ident!("{}", fn_name);
-
-        // Check if async
-        let is_async = n.function.is_async;
-
-        // Extract parameters
-        let mut params = Vec::new();
-        for param in &n.function.params {
-            if let Pat::Ident(ident_pat) = &param.pat {
-                let param_name = format_ident!("{}", ident_pat.sym.to_string());
-                let param_type = map_ts_type(ident_pat.type_ann.as_ref());
-                params.push(quote! { #param_name: #param_type });
-            }
-        }
-
-        // Extract return type - unwrap Promise<T> for async functions
-        let return_type = if is_async {
-            if n.function.return_type.is_none() {
-                // If no return type, assume void for async functions
-                quote! { Result<(), crate::AppError> }
-            } else {
-                let inner = unwrap_promise_type(n.function.return_type.as_ref());
-                quote! { Result<#inner, crate::AppError> }
-            }
-        } else if n.function.return_type.is_none() {
-            quote! { () }
-        } else {
-            map_ts_type(n.function.return_type.as_ref())
-        };
-
-        // Check if void
-        let is_void = if n.function.return_type.is_none() {
-            true
-        } else {
-            super::type_mapper::is_void_or_promise_void(n.function.return_type.as_deref())
-        };
-
-        // Convert body
-        let mut body_stmts = Vec::new();
-        if let Some(block_stmt) = &n.function.body {
-            if is_async {
-                for stmt in &block_stmt.stmts {
-                    body_stmts.push(self.convert_stmt_recursive(stmt, &mut |ret_stmt| {
-                        if let Some(arg) = &ret_stmt.arg {
-                            let expr = self.convert_expr(arg);
-                            if !is_void && matches!(arg.as_ref(), swc_ecma_ast::Expr::Object(_)) {
-                                quote! {
-                                    return Ok(serde_json::from_value(#expr).unwrap_or_default());
-                                }
-                            } else {
-                                quote! { return Ok(#expr); }
-                            }
-                        } else {
-                            quote! { return Ok(()); }
-                        }
-                    }));
-                }
-            } else {
-                for stmt in &block_stmt.stmts {
-                    body_stmts.push(self.convert_stmt_recursive(stmt, &mut |ret_stmt| {
-                        if let Some(arg) = &ret_stmt.arg {
-                            let expr = self.convert_expr(arg);
-                            if !is_void && matches!(arg.as_ref(), swc_ecma_ast::Expr::Object(_)) {
-                                quote! {
-                                    return serde_json::from_value(#expr).unwrap_or_default();
-                                }
-                            } else {
-                                quote! { return #expr; }
-                            }
-                        } else {
-                            quote! { return; }
-                        }
-                    }));
-                }
-            }
-        }
-
-        let vis = if self.is_exporting {
-            quote! { pub }
-        } else {
-            quote! {}
-        };
-
-        let generics = if let Some(type_params) = &n.function.type_params {
-            let params: Vec<_> = type_params
-                .params
-                .iter()
-                .map(|p| {
-                    let name = p.name.sym.to_string();
-                    let ident = format_ident!("{}", name);
-                    quote! { #ident: serde::de::DeserializeOwned + serde::Serialize + Clone }
-                })
-                .collect();
-            quote! { <#(#params),*> }
-        } else {
-            quote! {}
-        };
-
-        let fn_def = if is_async {
-            let fallback = if is_void {
-                quote! { Ok(()) }
-            } else {
-                quote! {}
-            };
-
-            quote! {
-                #vis async fn #fn_ident #generics (#(#params),*) -> #return_type {
-                    #(#body_stmts)*
-                    #fallback
-                }
-            }
-        } else {
-            quote! {
-                #vis fn #fn_ident #generics (#(#params),*) -> #return_type {
-                    #(#body_stmts)*
-                }
-            }
-        };
-
-        self.code.push_str(&fn_def.to_string());
-        self.code.push('\n');
-    }
-
-    // New helper method for recursive stmt conversion with callback
-    pub fn convert_stmt_recursive<F>(
-        &self,
-        stmt: &Stmt,
-        return_handler: &mut F,
-    ) -> proc_macro2::TokenStream
-    where
-        F: FnMut(&swc_ecma_ast::ReturnStmt) -> proc_macro2::TokenStream,
-    {
-        match stmt {
-            Stmt::Return(ret_stmt) => return_handler(ret_stmt),
-            Stmt::Block(block) => {
-                let stmts: Vec<_> = block
-                    .stmts
-                    .iter()
-                    .map(|s| self.convert_stmt_recursive(s, return_handler))
-                    .collect();
-                quote! { { #(#stmts)* } }
-            }
-            Stmt::If(if_stmt) => {
-                let test = self.convert_expr(&if_stmt.test);
-                let cons = self.convert_stmt_recursive(&if_stmt.cons, return_handler);
-                let cons_block = if matches!(*if_stmt.cons, Stmt::Block(_)) {
-                    quote! { #cons }
-                } else {
-                    quote! { { #cons } }
-                };
-
-                let alt = if let Some(alt) = &if_stmt.alt {
-                    let alt_stmt = self.convert_stmt_recursive(alt, return_handler);
-                    let alt_block = if matches!(&**alt, Stmt::Block(_) | Stmt::If(_)) {
-                        quote! { #alt_stmt }
-                    } else {
-                        quote! { { #alt_stmt } }
-                    };
-                    quote! { else #alt_block }
-                } else {
-                    quote! {}
-                };
-
-                quote! { if #test #cons_block #alt }
-            }
-            _ => self.convert_stmt(stmt),
-        }
-    }
-
-    pub fn convert_stmt(&self, stmt: &Stmt) -> proc_macro2::TokenStream {
-        match stmt {
-            Stmt::Return(ret_stmt) => {
-                if let Some(arg) = &ret_stmt.arg {
-                    let expr = self.convert_expr(arg);
-                    quote! { return #expr; }
-                } else {
-                    quote! { return; }
-                }
-            }
-            Stmt::Expr(expr_stmt) => {
-                let expr = self.convert_expr(&expr_stmt.expr);
-                quote! { #expr; }
-            }
-            Stmt::Decl(Decl::Var(var_decl)) => {
-                let mut declarations = Vec::new();
-                for decl in &var_decl.decls {
-                    let init_expr_opt = decl.init.as_ref().map(|init| self.convert_expr(init));
-
-                    match &decl.name {
-                        Pat::Ident(ident) => {
-                            let var_name = to_snake_case(&ident.id.sym);
-                            let var_ident = format_ident!("{}", var_name);
-
-                            if let Some(init_expr) = init_expr_opt {
-                                // Always use `let mut` for both let and const to support Rust's strict mutability rules unless optimization proves otherwise
-                                declarations.push(quote! {
-                                    let mut #var_ident = #init_expr;
-                                });
-                            } else {
-                                declarations.push(quote! {
-                                    let mut #var_ident;
-                                });
-                            }
-                        }
-                        Pat::Object(obj_pat) => {
-                            // Object destructuring: const { id, name = "default" } = expr
-                            if let Some(init_expr) = &init_expr_opt {
-                                let source_ident = format_ident!("__destructured");
-                                declarations.push(quote! {
-                                    let #source_ident = #init_expr;
-                                });
-                                for prop in &obj_pat.props {
-                                    match prop {
-                                        swc_ecma_ast::ObjectPatProp::KeyValue(kv) => {
-                                            if let swc_ecma_ast::PropName::Ident(key) = &kv.key {
-                                                let key_name = format_ident!(
-                                                    "{}",
-                                                    to_snake_case(key.sym.as_ref())
-                                                );
-                                                if let Pat::Ident(val_ident) = &*kv.value {
-                                                    let val_name = format_ident!(
-                                                        "{}",
-                                                        to_snake_case(val_ident.sym.as_ref())
-                                                    );
-                                                    declarations.push(quote! {
-                                                        let mut #val_name = #source_ident.#key_name.clone();
-                                                    });
-                                                }
-                                            }
-                                        }
-                                        swc_ecma_ast::ObjectPatProp::Assign(assign) => {
-                                            let key_name = format_ident!(
-                                                "{}",
-                                                to_snake_case(assign.key.sym.as_ref())
-                                            );
-                                            if let Some(default_val) = &assign.value {
-                                                let default_expr = self.convert_expr(default_val);
-                                                declarations.push(quote! {
-                                                    let mut #key_name = #source_ident.#key_name.clone().unwrap_or(#default_expr);
-                                                });
-                                            } else {
-                                                declarations.push(quote! {
-                                                    let mut #key_name = #source_ident.#key_name.clone();
-                                                });
-                                            }
-                                        }
-                                        swc_ecma_ast::ObjectPatProp::Rest(_) => {
-                                            declarations.push(
-                                                quote! { /* rest patterns not yet supported */ },
-                                            );
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        Pat::Array(arr_pat) => {
-                            // Array destructuring: const [first, second] = list
-                            if let Some(init_expr) = &init_expr_opt {
-                                let source_ident = format_ident!("__arr_destructured");
-                                declarations.push(quote! {
-                                    let #source_ident = #init_expr;
-                                });
-                                for (idx, elem) in arr_pat.elems.iter().enumerate() {
-                                    if let Some(Pat::Ident(ident)) = elem {
-                                        let var_name =
-                                            format_ident!("{}", to_snake_case(ident.sym.as_ref()));
-                                        let index = idx;
-                                        declarations.push(quote! {
-                                            let mut #var_name = #source_ident[#index].clone();
-                                        });
-                                    }
-                                }
-                            }
-                        }
-                        _ => {
-                            declarations.push(quote! { /* unsupported pattern */ });
-                        }
-                    }
-                }
-                quote! { #(#declarations)* }
-            }
-            Stmt::Block(block) => {
-                let stmts: Vec<_> = block.stmts.iter().map(|s| self.convert_stmt(s)).collect();
-                quote! { { #(#stmts)* } }
-            }
-            Stmt::If(if_stmt) => {
-                let test = self.convert_expr(&if_stmt.test);
-                let cons = self.convert_stmt(&if_stmt.cons);
-                let cons_block = if matches!(*if_stmt.cons, Stmt::Block(_)) {
-                    quote! { #cons }
-                } else {
-                    quote! { { #cons } }
-                };
-
-                let alt = if let Some(alt) = &if_stmt.alt {
-                    let alt_stmt = self.convert_stmt(alt);
-                    let alt_block = if matches!(&**alt, Stmt::Block(_) | Stmt::If(_)) {
-                        quote! { #alt_stmt }
-                    } else {
-                        quote! { { #alt_stmt } }
-                    };
-                    quote! { else #alt_block }
-                } else {
-                    quote! {}
-                };
-
-                quote! { if #test #cons_block #alt }
-            }
-            Stmt::While(while_stmt) => {
-                let test = self.convert_expr(&while_stmt.test);
-                let body = self.convert_stmt(&while_stmt.body);
-                let body_block = if matches!(*while_stmt.body, Stmt::Block(_)) {
-                    quote! { #body }
-                } else {
-                    quote! { { #body } }
-                };
-                quote! {
-                    while #test #body_block
-                }
-            }
-            Stmt::ForOf(for_of) => {
-                let body = self.convert_stmt(&for_of.body);
-                let right = self.convert_expr(&for_of.right);
-                let body_block = if matches!(*for_of.body, Stmt::Block(_)) {
-                    quote! { #body }
-                } else {
-                    quote! { { #body } }
-                };
-
-                // Extract the loop variable name from the declaration
-                let var_ident = match &for_of.left {
-                    swc_ecma_ast::ForHead::VarDecl(var_decl) => {
-                        if let Some(decl) = var_decl.decls.first() {
-                            if let Pat::Ident(ident) = &decl.name {
-                                let name = to_snake_case(&ident.id.sym);
-                                format_ident!("{}", name)
-                            } else {
-                                format_ident!("_item")
-                            }
-                        } else {
-                            format_ident!("_item")
-                        }
-                    }
-                    swc_ecma_ast::ForHead::Pat(pat) => {
-                        if let Pat::Ident(ident) = pat.as_ref() {
-                            let name = to_snake_case(&ident.id.sym);
-                            format_ident!("{}", name)
-                        } else {
-                            format_ident!("_item")
-                        }
-                    }
-                    _ => format_ident!("_item"),
-                };
-
-                quote! {
-                    for #var_ident in #right #body_block
-                }
-            }
-            Stmt::Throw(throw_stmt) => {
-                let arg = self.convert_expr(&throw_stmt.arg);
-                quote! {
-                    return Err(#arg.into());
-                }
-            }
-            _ => quote! { /* other stmts todo */ },
-        }
-    }
-
-    pub fn convert_expr(&self, expr: &Expr) -> proc_macro2::TokenStream {
-        match expr {
-            Expr::Bin(bin) => self.convert_bin_expr(bin),
-            Expr::This(_) => quote! { self },
-            Expr::Ident(ident) => {
-                let name = ident.sym.as_str();
-                if name == "undefined" {
-                    return quote! { None };
-                }
-                if name.chars().next().is_some_and(char::is_uppercase) {
-                    let ident_token = format_ident!("{}", name);
-                    quote! { #ident_token }
-                } else {
-                    let ident_name = to_snake_case(name);
-                    let ident_token = format_ident!("{}", ident_name);
-                    quote! { #ident_token }
-                }
-            }
-            Expr::Lit(lit) => match lit {
-                Lit::Num(num) => {
-                    let value = num.value;
-                    quote! { #value }
-                }
-                Lit::Str(s) => {
-                    let v = s.value.as_str().unwrap_or("");
-                    quote! { String::from(#v) }
-                }
-                Lit::Bool(b) => {
-                    let v = b.value;
-                    quote! { #v }
-                }
-                Lit::Null(_) => quote! { None },
-                _ => quote! { compile_error!("Tyrus: unsupported literal type") },
-            },
-            Expr::Member(member) => self.convert_member_expr(member),
-            Expr::Call(call) => self.convert_call_expr(call),
-            Expr::Object(obj) => self.convert_object_lit(obj),
-            Expr::Assign(assign) => self.convert_assign_expr(assign),
-            Expr::Update(update) => self.convert_update_expr(update),
-            Expr::Await(await_expr) => {
-                let arg = self.convert_expr(&await_expr.arg);
-                quote! { #arg.await? }
-            }
-            Expr::New(new_expr) => {
-                let callee = self.convert_expr(&new_expr.callee);
-                let args = if let Some(args) = &new_expr.args {
-                    args.iter()
-                        .map(|a| self.convert_expr_or_spread(a))
-                        .collect()
-                } else {
-                    Vec::new()
-                };
-                quote! { #callee::new(#(#args),*) }
-            }
-            Expr::Paren(paren) => self.convert_expr(&paren.expr),
-            Expr::Arrow(arrow) => self.convert_arrow_expr(arrow),
-
-            Expr::Array(arr) => self.convert_array_lit(arr),
-            Expr::Tpl(tpl) => self.convert_tpl(tpl),
-            Expr::Cond(cond) => {
-                let test = self.convert_expr(&cond.test);
-                let cons = self.convert_expr(&cond.cons);
-                let alt = self.convert_expr(&cond.alt);
-                quote! { if #test { #cons } else { #alt } }
-            }
-            Expr::OptChain(opt_chain) => self.convert_opt_chain(opt_chain),
-            _ => quote! { compile_error!("Tyrus: unsupported expression") },
-        }
-    }
-
-    /// Convert optional chaining expression (e.g., `user.config?.theme`)
-    /// Maps to Rust `Option::and_then(|x| x.field)` or `.map(|x| x.field)` chains.
-    fn convert_opt_chain(
-        &self,
-        opt_chain: &swc_ecma_ast::OptChainExpr,
-    ) -> proc_macro2::TokenStream {
-        match &*opt_chain.base {
-            swc_ecma_ast::OptChainBase::Member(member) => {
-                let obj = self.convert_expr(&member.obj);
-                if let Some(prop_ident) = member.prop.as_ident() {
-                    let prop_name = format_ident!("{}", to_snake_case(prop_ident.sym.as_ref()));
-                    // obj?.prop → obj.as_ref().and_then(|__v| __v.prop.clone())
-                    // If obj is already Option, use and_then; otherwise just access
-                    quote! { #obj.as_ref().and_then(|__v| __v.#prop_name.clone()) }
-                } else {
-                    quote! { #obj }
-                }
-            }
-            swc_ecma_ast::OptChainBase::Call(call) => {
-                let callee = self.convert_expr(&call.callee);
-                let args: Vec<_> = call
-                    .args
-                    .iter()
-                    .map(|a| self.convert_expr(&a.expr))
-                    .collect();
-                quote! { #callee(#(#args),*) }
-            }
-        }
-    }
-
-    fn convert_tpl(&self, tpl: &swc_ecma_ast::Tpl) -> proc_macro2::TokenStream {
-        let mut fmt_str = String::new();
-        let mut args = Vec::new();
-
-        for (i, quasi) in tpl.quasis.iter().enumerate() {
-            if let Some(cooked) = &quasi.cooked {
-                fmt_str.push_str(cooked.as_str().unwrap_or(quasi.raw.as_str()));
-            } else {
-                fmt_str.push_str(quasi.raw.as_str());
-            }
-
-            if i < tpl.exprs.len() {
-                fmt_str.push_str("{}");
-                args.push(self.convert_expr(&tpl.exprs[i]));
-            }
-        }
-
-        quote! { format!(#fmt_str, #(#args),*) }
-    }
-
-    fn convert_array_lit(&self, arr: &swc_ecma_ast::ArrayLit) -> proc_macro2::TokenStream {
-        let elems: Vec<_> = arr
-            .elems
-            .iter()
-            .flatten()
-            .map(|elem| self.convert_expr_or_spread(elem))
-            .collect();
-        quote! { vec![#(#elems),*] }
-    }
-
-    pub fn convert_expr_or_spread(&self, arg: &ExprOrSpread) -> proc_macro2::TokenStream {
-        self.convert_expr(&arg.expr)
-    }
-
-    fn convert_member_expr(&self, member: &MemberExpr) -> proc_macro2::TokenStream {
-        if member.obj.is_this() {
-            if let Some(prop_ident) = member.prop.as_ident() {
-                let prop_name = to_snake_case(prop_ident.sym.as_ref());
-                let field_ident = format_ident!("{}", prop_name);
-
-                // CHECK STATE FIELDS
-                if let Some(type_str) = self.current_class_state_fields.get(prop_ident.sym.as_ref())
-                {
-                    let needs_deref = matches!(
-                        type_str.as_str(),
-                        "f64" | "bool" | "i32" | "usize" | "u64" | "i64"
-                    );
-
-                    if needs_deref {
-                        return quote! { *self.#field_ident.lock().unwrap_or_else(|e| e.into_inner()) };
-                    } else if type_str == "String" {
-                        // String is not Copy, so we must clone it to return/use as value from MutexGuard
-                        return quote! { self.#field_ident.lock().unwrap_or_else(|e| e.into_inner()).clone() };
-                    } else {
-                        // For non-primitives (like Vec), return the Guard (or ref?)
-                        // If we return *guard, we move out of mutex? No, implementation of Deref.
-                        // But accessing `self.users.lock().unwrap_or_else(|e| e.into_inner())` returns `MutexGuard`.
-                        // If we want to call methods on it (`.push()`), we need the Guard.
-                        // If we want to pass it to a function expecting `Vec`, we might need `&*` or `.clone()`.
-                        // But here we return TokenStream.
-                        return quote! { self.#field_ident.lock().unwrap_or_else(|e| e.into_inner()) };
-                    }
-                } else {
-                    // Use clone() to be safe for generic and non-Copy types
-                    return quote! { self.#field_ident.clone() };
-                }
-            }
-        }
-
-        let obj = self.convert_expr(&member.obj);
-        match &member.prop {
-            swc_ecma_ast::MemberProp::Ident(ident) => {
-                let name = ident.sym.as_ref();
-                // If it starts with uppercase, preserve it (Enum variant, etc.)
-                let prop_name = if name.chars().next().is_some_and(char::is_uppercase) {
-                    name.to_string()
-                } else {
-                    to_snake_case(name)
-                };
-                let prop = format_ident!("{}", prop_name);
-
-                // Fix for Enum access: if obj is Capitalized (e.g. Status), use ::
-                // This is a heuristic. Ideally we check the type map.
-                let obj_str = obj.to_string();
-                // Check if obj_str looks like a type name (Capitalized, no dots/method calls yet)
-                // Note: obj is a TokenStream, to_string gives "Status" or "self . status".
-                // Simple heuristic: if it's a single word starting with Uppercase, treat as Enum/Static.
-                if obj_str.chars().next().is_some_and(char::is_uppercase)
-                    && !obj_str.contains('.')
-                    && !obj_str.contains('(')
-                {
-                    return quote! { #obj::#prop };
-                }
-
-                quote! { #obj.#prop }
-            }
-            swc_ecma_ast::MemberProp::Computed(computed) => {
-                // Unwrap parens helper
-                let mut expr = &*computed.expr;
-                while let Expr::Paren(p) = expr {
-                    expr = &p.expr;
-                }
-
-                // If the index is a number literal, it must be usize for Vec/Array indexing
-                if let Expr::Lit(Lit::Num(n)) = expr {
-                    let idx = n.value as usize;
-                    quote! { #obj[#idx] }
-                } else {
-                    let expr_tokens = self.convert_expr(&computed.expr);
-                    // Note: If 'expr' is a variable 'i' (f64), we might need casting 'as usize'
-                    // But determining if #obj is a Vec vs Map is hard without type info.
-                    // For now, we fix the literal 0 case which is most common in tuples/split.
-                    quote! { #obj[#expr_tokens] }
-                }
-            }
-            _ => quote! { compile_error!("Tyrus: unsupported member expression") },
-        }
-    }
-
-    fn convert_bin_expr(&self, bin: &BinExpr) -> proc_macro2::TokenStream {
-        let left = self.convert_expr(&bin.left);
-        let right = self.convert_expr(&bin.right);
-        match bin.op {
-            BinaryOp::EqEq | BinaryOp::EqEqEq => quote! { #left == #right },
-            BinaryOp::NotEq | BinaryOp::NotEqEq => quote! { #left != #right },
-            BinaryOp::Add => {
-                if Self::is_string_expr(&bin.left) || Self::is_string_expr(&bin.right) {
-                    quote! { format!("{}{}", #left, #right) }
-                } else {
-                    quote! { #left + #right }
-                }
-            }
-            BinaryOp::Sub => quote! { #left - #right },
-            BinaryOp::Mul => quote! { #left * #right },
-            BinaryOp::Div => quote! { #left / #right },
-            BinaryOp::Mod => quote! { #left % #right },
-            BinaryOp::Lt => quote! { #left < #right },
-            BinaryOp::LtEq => quote! { #left <= #right },
-            BinaryOp::Gt => quote! { #left > #right },
-            BinaryOp::GtEq => quote! { #left >= #right },
-            BinaryOp::LogicalOr => quote! { #left || #right },
-            BinaryOp::LogicalAnd => quote! { #left && #right },
-            BinaryOp::NullishCoalescing => quote! { #left.unwrap_or(#right) },
-            _ => {
-                let op_str = format!("{:?}", bin.op);
-                quote! { compile_error!(concat!("Tyrus: unsupported binary operator: ", #op_str)) }
-            }
-        }
-    }
-
-    /// Heuristic to detect if an expression is likely a string type in TypeScript.
-    /// Used to choose `format!` over `+` for string concatenation.
-    fn is_string_expr(expr: &Expr) -> bool {
-        match expr {
-            Expr::Lit(Lit::Str(_)) => true,
-            Expr::Tpl(_) => true,
-            Expr::Call(call) => {
-                if let Callee::Expr(callee) = &call.callee {
-                    if let Expr::Member(member) = &**callee {
-                        if let Some(ident) = member.prop.as_ident() {
-                            return matches!(
-                                ident.sym.as_str(),
-                                "toString"
-                                    | "toUpperCase"
-                                    | "toLowerCase"
-                                    | "trim"
-                                    | "replace"
-                                    | "join"
-                                    | "slice"
-                                    | "substring"
-                                    | "charAt"
-                                    | "concat"
-                                    | "padStart"
-                                    | "padEnd"
-                                    | "repeat"
-                            );
-                        }
-                    }
-                    // String::from(...) / String(...)
-                    if let Expr::Ident(ident) = &**callee {
-                        return ident.sym.as_str() == "String";
-                    }
-                }
-                false
-            }
-            Expr::Bin(bin) if bin.op == BinaryOp::Add => {
-                Self::is_string_expr(&bin.left) || Self::is_string_expr(&bin.right)
-            }
-            _ => false,
-        }
-    }
-
-    fn convert_call_expr(&self, call: &CallExpr) -> proc_macro2::TokenStream {
-        // Try stdlib handlers first - PASS SELF
-        if let Some(stdlib_code) =
-            crate::stdlib::try_handle_stdlib_call(self, &call.callee, &call.args)
-        {
-            return stdlib_code;
-        }
-
-        // Check for axios calls (axios.get, axios.post, etc.)
-        if let Callee::Expr(expr) = &call.callee {
-            if let Expr::Member(member) = &**expr {
-                // Handle axios.get/post/put/delete → reqwest
-                if let Expr::Ident(obj_ident) = &*member.obj {
-                    if obj_ident.sym.as_str() == "axios" {
-                        if let Some(method_ident) = member.prop.as_ident() {
-                            let method_name = method_ident.sym.as_str();
-                            if matches!(method_name, "get" | "post" | "put" | "delete" | "patch") {
-                                let url = call
-                                    .args
-                                    .first()
-                                    .map(|a| self.convert_expr(&a.expr))
-                                    .unwrap_or_else(|| quote! { "" });
-                                let method_fn = format_ident!("{}", method_name);
-                                return quote! { reqwest::Client::new().#method_fn(#url).send() };
-                            }
-                        }
-                    }
-                }
-
-                // Try stdlib method call - PASS SELF
-                if let Some(method_ident) = member.prop.as_ident() {
-                    let method_name = method_ident.sym.as_ref();
-                    if let Some(stdlib_code) = crate::stdlib::try_handle_method_call(
-                        self,
-                        &member.obj,
-                        method_name,
-                        &call.args,
-                    ) {
-                        return stdlib_code;
-                    }
-                }
-            }
-        }
-
-        // Handle fetch() → reqwest::Client::new().method(url)
-        if let Callee::Expr(expr) = &call.callee {
-            if let Expr::Ident(ident) = &**expr {
-                if ident.sym.as_str() == "fetch" {
-                    let url = call
-                        .args
-                        .first()
-                        .map(|a| self.convert_expr(&a.expr))
-                        .unwrap_or_else(|| quote! { "" });
-                    // Check for options object with method
-                    let method = call.args.get(1).and_then(|opts| {
-                        if let Expr::Object(obj) = &*opts.expr {
-                            for prop in &obj.props {
-                                if let swc_ecma_ast::PropOrSpread::Prop(p) = prop {
-                                    if let swc_ecma_ast::Prop::KeyValue(kv) = &**p {
-                                        if let swc_ecma_ast::PropName::Ident(id) = &kv.key {
-                                            if id.sym.as_str() == "method" {
-                                                if let Expr::Lit(Lit::Str(s)) = &*kv.value {
-                                                    return Some(
-                                                        s.value
-                                                            .as_str()
-                                                            .unwrap_or_default()
-                                                            .to_uppercase(),
-                                                    );
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        None
-                    });
-                    let method_call = match method.as_deref() {
-                        Some("POST") => quote! { post },
-                        Some("PUT") => quote! { put },
-                        Some("DELETE") => quote! { delete },
-                        Some("PATCH") => quote! { patch },
-                        _ => quote! { get },
-                    };
-                    return quote! { reqwest::Client::new().#method_call(#url).send() };
-                }
-            }
-        }
-
-        let callee = match &call.callee {
-            Callee::Expr(expr) => {
-                // Intercept Array methods: map, filter, forEach, reduce, find, some, every
-                if let Expr::Member(member) = &**expr {
-                    if let Some(method_ident) = member.prop.as_ident() {
-                        let method_name = method_ident.sym.as_ref();
-                        match method_name {
-                            "map" | "filter" => {
-                                let obj = self.convert_expr(&member.obj);
-                                let method_ident = format_ident!("{}", to_snake_case(method_name));
-                                let args: Vec<_> = call
-                                    .args
-                                    .iter()
-                                    .map(|a| self.convert_expr(&a.expr))
-                                    .collect();
-
-                                // Check for 2-arg closure (item, index)
-                                let has_index_arg = if let Some(arg) = call.args.first() {
-                                    if let Expr::Arrow(arrow) = &*arg.expr {
-                                        arrow.params.len() == 2
-                                    } else if let Expr::Fn(fn_expr) = &*arg.expr {
-                                        fn_expr.function.params.len() == 2
-                                    } else {
-                                        false
-                                    }
-                                } else {
-                                    false
-                                };
-
-                                if has_index_arg {
-                                    let closure = &args[0];
-                                    if method_name == "filter" {
-                                        // filter: handle &(i, v) pattern and strip index from result
-                                        return quote! {
-                                            #obj.clone().into_iter()
-                                                .enumerate()
-                                                .filter(|(i, v)| (#closure)(v.clone(), *i as f64))
-                                                .map(|(_, v)| v)
-                                                .collect::<Vec<_>>()
-                                        };
-                                    } else {
-                                        // map: direct mapping
-                                        return quote! {
-                                            #obj.clone().into_iter()
-                                                .enumerate()
-                                                .map(|(i, v)| (#closure)(v, i as f64))
-                                                .collect::<Vec<_>>()
-                                        };
-                                    }
-                                }
-
-                                // Eager evaluation
-                                if method_name == "filter" {
-                                    // filter passes &T to closure; we inline with clone to avoid reference issues
-                                    if let Some(first_arg) = call.args.first() {
-                                        if let Expr::Arrow(arrow) = &*first_arg.expr {
-                                            if let Some(Pat::Ident(param_ident)) =
-                                                arrow.params.first()
-                                            {
-                                                let param_name = format_ident!(
-                                                    "{}",
-                                                    to_snake_case(param_ident.sym.as_ref())
-                                                );
-                                                let body = match &*arrow.body {
-                                                    swc_ecma_ast::BlockStmtOrExpr::Expr(expr) => {
-                                                        self.convert_expr(expr)
-                                                    }
-                                                    swc_ecma_ast::BlockStmtOrExpr::BlockStmt(
-                                                        block,
-                                                    ) => {
-                                                        let stmts: Vec<_> = block
-                                                            .stmts
-                                                            .iter()
-                                                            .map(|s| self.convert_stmt(s))
-                                                            .collect();
-                                                        quote! { #(#stmts)* }
-                                                    }
-                                                };
-                                                return quote! {
-                                                    #obj.clone().into_iter()
-                                                        .filter(|#param_name| { let #param_name = #param_name.clone(); #body })
-                                                        .collect::<Vec<_>>()
-                                                };
-                                            }
-                                        }
-                                    }
-                                    // Fallback: use the pre-converted closure as-is
-                                    let closure = &args[0];
-                                    return quote! {
-                                        #obj.clone().into_iter()
-                                            .filter(|__v| (#closure)(__v.clone()))
-                                            .collect::<Vec<_>>()
-                                    };
-                                }
-                                return quote! { #obj.clone().into_iter().#method_ident(#(#args),*).collect::<Vec<_>>() };
-                            }
-                            "forEach" => {
-                                let obj = self.convert_expr(&member.obj);
-                                let args: Vec<_> = call
-                                    .args
-                                    .iter()
-                                    .map(|a| self.convert_expr(&a.expr))
-                                    .collect();
-
-                                let has_index_arg = if let Some(arg) = call.args.first() {
-                                    if let Expr::Arrow(arrow) = &*arg.expr {
-                                        arrow.params.len() == 2
-                                    } else if let Expr::Fn(fn_expr) = &*arg.expr {
-                                        fn_expr.function.params.len() == 2
-                                    } else {
-                                        false
-                                    }
-                                } else {
-                                    false
-                                };
-
-                                if has_index_arg {
-                                    let closure = &args[0];
-                                    return quote! {
-                                        #obj.into_iter()
-                                            .enumerate()
-                                            .for_each(|(i, v)| (#closure)(v, i as f64))
-                                    };
-                                }
-
-                                return quote! { #obj.into_iter().for_each(#(#args),*) };
-                            }
-                            "some" => {
-                                let obj = self.convert_expr(&member.obj);
-                                let args: Vec<_> = call
-                                    .args
-                                    .iter()
-                                    .map(|a| self.convert_expr(&a.expr))
-                                    .collect();
-                                return quote! { #obj.into_iter().any(#(#args),*) };
-                            }
-                            "every" => {
-                                let obj = self.convert_expr(&member.obj);
-                                let args: Vec<_> = call
-                                    .args
-                                    .iter()
-                                    .map(|a| self.convert_expr(&a.expr))
-                                    .collect();
-                                return quote! { #obj.into_iter().all(#(#args),*) };
-                            }
-                            "find" | "reduce" => {
-                                let obj = self.convert_expr(&member.obj);
-                                let method_ident = format_ident!("{}", to_snake_case(method_name));
-                                let args: Vec<_> = call
-                                    .args
-                                    .iter()
-                                    .map(|a| self.convert_expr(&a.expr))
-                                    .collect();
-                                return quote! { #obj.into_iter().#method_ident(#(#args),*) };
-                            }
-                            "push" => {
-                                let method_ident = format_ident!("push");
-                                let args: Vec<_> = call
-                                    .args
-                                    .iter()
-                                    .map(|a| self.convert_expr(&a.expr))
-                                    .collect();
-
-                                // Check for this.field.push()
-                                // member.obj is the expression on which push is called (e.g. this.field)
-                                if let Expr::Member(nested_member) = &*member.obj {
-                                    if nested_member.obj.is_this() {
-                                        if let Some(prop_ident) = nested_member.prop.as_ident() {
-                                            let prop_name = to_snake_case(prop_ident.sym.as_ref());
-                                            let field = format_ident!("{}", prop_name);
-                                            // Access field directly for mutation
-                                            return quote! { self.#field.#method_ident(#(#args),*) };
-                                        }
-                                    }
-                                }
-
-                                if member.obj.is_this() {
-                                    // this.push() - rare
-                                    return quote! { self.#method_ident(#(#args),*) };
-                                }
-
-                                let obj = self.convert_expr(&member.obj);
-                                return quote! { #obj.#method_ident(#(#args),*) };
-                            }
-                            "replace" => {
-                                let method_ident = format_ident!("replacen");
-                                let args: Vec<_> = call
-                                    .args
-                                    .iter()
-                                    .map(|a| self.convert_expr(&a.expr))
-                                    .collect();
-
-                                let obj = self.convert_expr(&member.obj);
-                                // TS replace(str, str) only replaces first occurrence. Rust replace() replaces all.
-                                // We use replacen(pat, to, 1).
-                                if args.len() == 2 {
-                                    return quote! { #obj.#method_ident(#(#args),*, 1) };
-                                }
-                                // Fallback for other signatures (regex?) - though replacen expects 3 args
-                                return quote! { #obj.replace(#(#args),*) };
-                            }
-                            _ => {}
-                        }
-                    }
-                }
-                self.convert_expr(expr)
-            }
-            _ => quote! { compile_error!("Tyrus: unsupported call expression") },
-        };
-        let args: Vec<_> = call
-            .args
-            .iter()
-            .map(|a| self.convert_expr(&a.expr))
-            .collect();
-        quote! { #callee(#(#args),*) }
-    }
-
-    fn convert_object_lit(&self, obj: &swc_ecma_ast::ObjectLit) -> proc_macro2::TokenStream {
-        let mut fields = Vec::new();
-        for prop in &obj.props {
-            if let swc_ecma_ast::PropOrSpread::Prop(p) = prop {
-                if let swc_ecma_ast::Prop::KeyValue(kv) = &**p {
-                    if let swc_ecma_ast::PropName::Ident(ident) = &kv.key {
-                        let key = ident.sym.as_ref();
-
-                        // Special handling for null/undefined in JSON macro context
-                        // They must be explicit Value::Null to avoid inference errors with None
-                        let is_null = matches!(&*kv.value, Expr::Lit(Lit::Null(_)));
-                        let is_undefined = if let Expr::Ident(id) = &*kv.value {
-                            id.sym.as_ref() == "undefined"
-                        } else {
-                            false
-                        };
-
-                        if is_null || is_undefined {
-                            fields.push(quote! { #key: serde_json::Value::Null });
-                        } else {
-                            let val = self.convert_expr(&kv.value);
-                            fields.push(quote! { #key: #val });
-                        }
-                    }
-                }
-            }
-        }
-        quote! { serde_json::json!({ #(#fields),* }) }
-    }
-
-    fn convert_assign_expr(&self, assign: &swc_ecma_ast::AssignExpr) -> proc_macro2::TokenStream {
-        let right = self.convert_expr(&assign.right);
-        // Handle Left Side
-        let left = match &assign.left {
-            swc_ecma_ast::AssignTarget::Simple(simple) => {
-                match simple {
-                    swc_ecma_ast::SimpleAssignTarget::Member(member) => {
-                        if member.obj.is_this() {
-                            if let Some(prop_ident) = member.prop.as_ident() {
-                                let prop_name = to_snake_case(prop_ident.sym.as_ref());
-                                let field = format_ident!("{}", prop_name);
-
-                                // CHECK STATE FIELDS
-                                if let Some(_type_str) =
-                                    self.current_class_state_fields.get(prop_ident.sym.as_ref())
-                                {
-                                    // Primitive check?
-                                    // f64, bool, i32, usize etc. need dereference when used as value, but wait...
-                                    // In Assign Expr:
-                                    // LHS = RHS
-                                    // If LHS is Mutex, we need `*lock() = value`.
-                                    // This applies to ALL types inside Mutex if we are replacing the value?
-                                    // No, only if we can assign to deref.
-                                    // `Vec` inside Mutex: `*lock() = new_vec`. Yes.
-
-                                    // So actually, ALL assignments to Mutex-wrapped field need `*`?
-                                    // `*self.users.lock().unwrap() = ...`
-                                    // Yes, otherwise we are trying to assign to `MutexGuard` temporary?
-                                    // Assigning to `*guard` assigns to the inner value.
-
-                                    quote! { *self.#field.lock().unwrap_or_else(|e| e.into_inner()) }
-                                } else {
-                                    quote! { self.#field }
-                                }
-                            } else {
-                                quote! { compile_error!("Tyrus: unsupported self member assignment") }
-                            }
-                        } else if let Some(prop_ident) = member.prop.as_ident() {
-                            let prop_name = to_snake_case(prop_ident.sym.as_ref());
-                            let field = format_ident!("{}", prop_name);
-                            let obj = self.convert_expr(&member.obj);
-                            quote! { #obj.#field }
-                        } else {
-                            quote! { compile_error!("Tyrus: unsupported member assignment pattern") }
-                        }
-                    }
-                    swc_ecma_ast::SimpleAssignTarget::Ident(ident) => {
-                        let name = format_ident!("{}", to_snake_case(ident.sym.as_ref()));
-                        quote! { #name }
-                    }
-                    _ => quote! { compile_error!("Tyrus: unsupported assignment target") },
-                }
-            }
-            _ => quote! { compile_error!("Tyrus: unsupported assignment pattern") },
-        };
-
-        match assign.op {
-            swc_ecma_ast::AssignOp::Assign => quote! { #left = #right },
-            swc_ecma_ast::AssignOp::AddAssign => quote! { #left += #right },
-            _ => quote! { #left op= #right },
-        }
-    }
-
-    fn convert_update_expr(&self, update: &UpdateExpr) -> proc_macro2::TokenStream {
-        let arg = self.convert_expr(&update.arg);
-        // If arg is a mutex lock result, modifications work assuming deref mut
-        match update.op {
-            UpdateOp::PlusPlus => quote! { #arg += 1.0 },
-            UpdateOp::MinusMinus => quote! { #arg -= 1.0 },
-        }
-    }
-
-    fn convert_arrow_expr(&self, arrow: &ArrowExpr) -> proc_macro2::TokenStream {
-        let params: Vec<_> = arrow
-            .params
-            .iter()
-            .map(|pat| {
-                if let Pat::Ident(ident) = pat {
-                    let name = format_ident!("{}", to_snake_case(&ident.id.sym));
-                    quote! { #name }
-                } else {
-                    quote! { _ }
-                }
-            })
-            .collect();
-
-        let body = match &*arrow.body {
-            BlockStmtOrExpr::BlockStmt(block) => {
-                let stmts: Vec<_> = block.stmts.iter().map(|s| self.convert_stmt(s)).collect();
-                quote! { { #(#stmts)* } }
-            }
-            BlockStmtOrExpr::Expr(expr) => self.convert_expr(expr),
-        };
-
-        // Note: We might need type annotations for params in complex cases,
-        // but for simple .find(|x| ...), inference usually works.
-        quote! {
-            |#(#params),*| #body
-        }
-    }
-}
